Creation of the Blandeu Portfolio project
Creation of the page 'welcome.html' that welcomes you to the website, and which redirects you to 'home.html' in the page directory. Added the 3 layers of my logo to the img directory to animate it in due time. 'script.js' & 'styles.css' that will define the core functions and appearance of the website.